Calculate quantized convolution multiplier using double type.
NNAPI has been using the float type for quantized convolution multiplier
calculation, which was causing an occasional off-by-one differences for
convolution operation results between TFLite/CPU and NNAPI/CPU implementations.

Test: MLTS com.android.nn.benchmark.app.NNTest#testNNAPI[mobilenet_*_quant] error rates
Test: NeuralNetworksTest_static
Bug: 115377572
Merged-In: If0bb9ff7229f5b3864a69c66566a4e4008f94c37
(cherry picked from commit e43841a12f362c204cab5caacc2190fe0eeb8051)

+++ common/operations/Conv2D.cpp
@@ -199,7 +199,7 @@ bool convNhwc(const uint8_t* inputData, const Shape& inputShape, const uint8_t* 
     int32_t filterOffset = -filterShape.offset;
     int32_t outputOffset = outputShape.offset;
 
-    float real_multiplier = 0.0;
+    double real_multiplier = 0.0;
     int32_t output_multiplier = 0;
     int32_t output_shift = 0;
     int32_t output_activation_min = 0;
@@ -302,7 +302,7 @@ bool convQuant8PerChannelNhwc(const uint8_t* inputData, const Shape& inputShape,
     int32_t inputOffset = -inputShape.offset;
     int32_t outputOffset = outputShape.offset;
 
-    auto realMultiplier = std::vector<float>(outputDepth, .0f);
+    auto realMultiplier = std::vector<double>(outputDepth, .0f);
     auto outputMultiplier = std::vector<int32_t>(outputDepth, 0);
     auto outputShift = std::vector<int32_t>(outputDepth, .0f);
